Split opcode tests
